refactor(shared): move base64url helpers to the canonical base64 module
Move toBase64Url/fromBase64Url from shared/identity-header.ts into
shared/base64.ts alongside their standard-base64 siblings, and export
identity-header's isByteStringSafe instead of leaving it private -
three test files were re-implementing it inline, one with a subtly
different (codePointAt-in-a-length-indexed-loop) iteration than the
real charCodeAt version. Also documents the deploy-ordering
constraint on decodeUserIdHeader's `=` sentinel for mixed-rollout
safety.
