+++ packages/runtime/__tests__/shard-client.test.ts
@@ -8,7 +8,7 @@
  */
 import { describe, expect, it } from "vitest";
 
-import { encodeIdentityHeader } from "../../../shared/identity-header";
+import { encodeIdentityHeader, isByteStringSafe } from "../../../shared/identity-header";
 import type { ShardNamespaceLike } from "../src/resolve-shard";
 import type { ShardFunctionReference } from "../src/shard-client";
 import { createShardClient } from "../src/shard-client";
@@ -119,18 +119,11 @@ describe("createShardClient", () => {
             .call("mcp:listNodes", {});
 
         const identityHeader = calls[0]?.headers["x-lunora-identity"] ?? "";
-        let isByteStringSafe = identityHeader.length > 0;
-
-        for (let index = 0; index < identityHeader.length; index += 1) {
-            if ((identityHeader.codePointAt(index) ?? 0) > 255) {
-                isByteStringSafe = false;
-            }
-        }
 
         // Every code unit must be <= 255 (WebIDL `ByteString`-safe) — this is the
         // literal property `new Request(...)` enforces at the real fetch call below;
         // a raw `JSON.stringify({ name: "名前 🎌" })` would violate it.
-        expect(isByteStringSafe).toBe(true);
+        expect(isByteStringSafe(identityHeader)).toBe(true);
         expect(identityHeader).toBe(encodeIdentityHeader({ name: "名前 🎌" }));
     });
