+++ shared/base64.ts
@@ -7,6 +7,10 @@
  * Both use a fixed-size chunk so a large buffer — a multi-megabyte audio utterance
  * or a `v.bytes()` payload — never overflows the argument-count ceiling of
  * `String.fromCharCode` (nor blows the call stack) via a single spread.
+ *
+ * `toBase64Url`/`fromBase64Url` are the URL-safe, unpadded variant built on top
+ * of the two above; `shared/identity-header.ts` imports them from here as the
+ * canonical base64url home rather than hand-rolling its own.
  */
 
 /** Base64-encode bytes. */
@@ -33,4 +37,15 @@ const fromBase64 = (base64: string): Uint8Array => {
     return bytes;
 };
 
-export { fromBase64, toBase64 };
+/** Base64 (from {@link toBase64}) -> URL-safe, unpadded base64url. */
+const toBase64Url = (bytes: Uint8Array): string => toBase64(bytes).replaceAll("+", "-").replaceAll("/", "_").replace(/=+$/, "");
+
+/** URL-safe base64url -> bytes, via {@link fromBase64} (which expects standard padded base64). */
+const fromBase64Url = (value: string): Uint8Array => {
+    const restored = value.replaceAll("-", "+").replaceAll("_", "/");
+    const padded = restored + "=".repeat((4 - (restored.length % 4)) % 4);
+
+    return fromBase64(padded);
+};
+
+export { fromBase64, fromBase64Url, toBase64, toBase64Url };
